+++ packages/scripts/src/commands/app-data/convert/processors/flowParser/parsers/data_pipe.parser.spec.ts
@@ -1,5 +1,5 @@
 import { FlowTypes } from "data-models";
-import { IDataPipeOperation } from "shared";
+import { IDataPipeOperation } from "shared/src/models/dataPipe/types";
 import { DataPipeParser } from "./data_pipe.parser";
 
 const testInputSources = {
